feat: 확인 글자가 삭제일 경우 텍스트 색상 빨간색으로 적용
