+++ styles.css
@@ -542,6 +542,10 @@ table {
   color: var(--danger-text-default);
 }
 
+.danger-btn {
+  color: var(--danger-text-default);
+}
+
 .delete-btn-icon {
   margin: 0 6px;
   color: var(--nuetral-surface-default);
@@ -653,7 +657,7 @@ dialog {
 .modal-btn {
   width: 100%;
   height: 100%;
-  border: none;
+  border-top: 1px solid var(--nuetral-border-default);
   outline: none;
   cursor: pointer;
 }
@@ -668,6 +672,10 @@ dialog {
   display: flex;
   justify-content: space-between;
   align-items: center;
+}
+
+.modal-btn:first-child {
+  border-right: 1px solid var(--nuetral-border-default);
 }
 
 .modal-content {
